rename tokenizer/pattern_to_token to base
