+++ uttut/pipeline/ops/pattern_recognizers/base.py
