+++ uttut/pipeline/ops/pattern_recognizers/float_token.py
@@ -2,7 +2,7 @@ import re
 
 from ..tokens import FLOAT_TOKEN
 from .int_token import IntTokenAligner
-from .pattern_to_token import PatternRecognizer
+from .base import PatternRecognizer
 
 
 class FloatToken(PatternRecognizer):
