+++ uttut/pipeline/ops/pattern_recognizers/int_token_with_space.py
@@ -6,7 +6,7 @@ from collections import Counter
 from uttut import ENTITY_LABEL
 from ..label_transducer import get_most_common_except_not_entity
 from ..tokens import INT_TOKEN_WITH_SPACE
-from .pattern_to_token import PatternRecognizer, PatternRecognizerAligner
+from .base import PatternRecognizer, PatternRecognizerAligner
 
 
 class IntTokenWithSpace(PatternRecognizer):
