+++ uttut/pipeline/ops/pattern_recognizers/merge_whitespace_characters.py
@@ -3,7 +3,7 @@ from typing import List
 import re
 
 from ..label_transducer import get_most_common
-from .pattern_to_token import PatternRecognizer, PatternRecognizerAligner
+from .base import PatternRecognizer, PatternRecognizerAligner
 
 
 class MergeWhiteSpaceCharacters(PatternRecognizer):
