+++ uttut/pipeline/ops/pattern_recognizers/strip_whitespace_characters.py
@@ -3,7 +3,7 @@ from typing import List
 import re
 
 from ..label_transducer import get_not_entity
-from .pattern_to_token import PatternRecognizer, PatternRecognizerAligner
+from .base import PatternRecognizer, PatternRecognizerAligner
 
 
 class StripWhiteSpaceCharacters(PatternRecognizer):
