+++ uttut/pipeline/ops/tokenizers/custom_word_tokenizer.py
@@ -2,7 +2,7 @@ from typing import List
 
 from ..label_transducer import get_most_common_except_not_entity
 from ..utils.trie import Trie
-from .tokenizer import Tokenizer, TokenizerAligner
+from .base import Tokenizer, TokenizerAligner
 
 
 class CustomWordTokenizer(Tokenizer):
