+++ uttut/pipeline/ops/tokenizers/eng_tokenizer.py
@@ -2,8 +2,8 @@ from typing import List
 
 import unicodedata
 
-from .tokenizer import Tokenizer, TokenizerAligner
 from ..label_transducer import get_most_common_except_not_entity
+from .base import Tokenizer, TokenizerAligner
 
 
 class EngTokenizer(Tokenizer):
